+++ apps/AppShell/.env.development
@@ -1,5 +1,5 @@
 ASSETS_PATH=/
 GENERATE_SOURCEMAP=true
 PORT=3000
-REMOTE_URLS={ "mfe_footer": "mfe_footer@http://localhost:3003", "mfe_header":"mfe_header@http://localhost:9000/localhost:3001", "mfe_product_listing":"mfe_product_listing@http://localhost:3005", "mfe_product_details":"mfe_product_details@http://localhost:3006" }
+REMOTE_URLS={ "mfe_footer": "mfe_footer@http://localhost:3003", "mfe_header":"mfe_header@http://localhost:3001", "mfe_product_listing":"mfe_product_listing@http://localhost:3005", "mfe_product_details":"mfe_product_details@http://localhost:3006" }
 LOG_LEVEL=info
